% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ufRisk.R
\name{plot.ufRisk}
\alias{plot.ufRisk}
\title{Plot Method for the Package 'ufRisk'}
\usage{
\method{plot}{ufRisk}(x, plot.option = NULL, ...)
}
\arguments{
\item{x}{an input object of class \code{ufRisk}.}

\item{plot.option}{plot choice for an object of class \code{ufRisk}; viable
choices are:
\describe{
\item{1}{Plotting out-of-sample loss series}
\item{2}{Plotting out-of-sample losses, VaR.v & breaches}
\item{3}{Plotting out-of-sample losses, VaR.e, ES & breaches}
}
Please note if no value is passed to \code{plot.option} a selection menu is
prompted; is set to \code{NULL} by default.}

\item{...}{additional arguments of the standard plot method.}
}
\value{
None
}
\description{
This function regulates how objects created by the package \code{ufRisk} are
plotted.
}
\author{
\itemize{
\item Sebastian Letmathe (Scientific Employee) (Department of Economics,
Paderborn University) \cr
\item Dominik Schulz (Research Assistant) (Department of Economics, Paderborn
University), \cr
}
}
