% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ufRisk.R
\name{print.ufRisk}
\alias{print.ufRisk}
\title{Print Method for Objects of Class 'ufRisk'}
\usage{
\method{print}{ufRisk}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ufRisk}; for the current package version,
only the function \code{br_test} returns such an object.}

\item{...}{implemented for compatibility with the generic function;
additional arguments, however, will not affect this print method.}
}
\value{
None
}
\description{
This function regulates how objects of class \code{ufRisk} are printed.
}
\author{
\itemize{
\item Sebastian Letmathe (Scientific Employee) (Department of Economics,
Paderborn University) \cr
\item Dominik Schulz (Scientific Employee) (Department of Economics,
Paderborn University), \cr
}
}
