% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WMT.R
\docType{data}
\name{WMT}
\alias{WMT}
\title{Walmart Inc. (WMT) Financial Time Series Data}
\format{
A data frame with 5535 rows and 10 variables:
\describe{
  \item{price.open}{opening price (daily)}
  \item{price.high}{highest price (daily)}
  \item{price.low}{lowest price (daily)}
  \item{price.close}{closing price (daily)}
  \item{volume}{trading volume}
  \item{price.adjusted}{adjusted closing price (daily)}
  \item{ref.date}{date in format YY-MM-DD}
  \item{ticker}{ticker symbol}
  \item{ret.adjusted.prices}{returns obtained from the adjusted closing
  prices}
  \item{ret.closing.prices}{returns obtained from the closing prices}
}
}
\source{
The data was obtained from Yahoo Finance.
}
\usage{
WMT
}
\description{
A dataset that contains the daily financial data of WMT from
January 2000 to December 2021 (currency in EUR).
}
\keyword{datasets}
